FEAT: 디렉토리 구조화
