+++ backend/src/main/java/com/teamk/swiths_api/global/Major.java
@@ -1,31 +1,24 @@
-package com.teamk.swiths_api.domain;
+package com.teamk.swiths_api.global;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
 @Getter
 @Setter
-public class UserClub {
+public class Major {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(unique = true, nullable = false)
     private Long id;
 
-    @ManyToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "club_id", nullable = false)
-    private Club club;
+    @Column(length = 50, nullable = false)
+    private String name; // 과목 명
 
 }
