+++ backend/src/main/java/com/teamk/swiths_api/post/Post.java
@@ -1,5 +1,9 @@
-package com.teamk.swiths_api.domain;
+package com.teamk.swiths_api.post;
 
+import jakarta.persistence.OneToOne;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,16 +11,17 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+
+import com.teamk.swiths_api.post.vote.Vote;
+import com.teamk.swiths_api.user.User;
 
 @Entity
 @Getter
 @Setter
-public class Comment {
+public class Post {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -24,15 +29,17 @@ public class Comment {
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "post_id")
-    private Post post;
+    @JoinColumn(name = "user_id", nullable = false)
+    public User user;
 
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
+    @Column(length = 20, nullable = false)
+    private String title; // 스터디명
 
-    @Column(name = "content", columnDefinition = "TEXT") // Mysql TEXT data type
-    private String content;
+    @Column(name = "short_content", length = 20)
+    private String shortContent; // 썸네일 상 보이는 간략 설명
+
+    @Column(nullable = false, columnDefinition = "TEXT")
+    private String content; // 스터디 설명창을 누르면 보이는 상세 설명 창
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -41,4 +48,9 @@ public class Comment {
     @UpdateTimestamp
     @Column(name = "updated_at")
     private LocalDateTime updated_at = LocalDateTime.now();
+
+    @OneToOne
+    @JoinColumn(name = "vote_id")
+    private Vote vote;
+
 }
