+++ backend/src/main/java/com/teamk/swiths_api/post/PostRepository.java
@@ -1,0 +1,5 @@
+package com.teamk.swiths_api.post;
+
+public interface PostRepository {
+    
+}
