+++ backend/src/main/java/com/teamk/swiths_api/post/PostService.java
@@ -1,0 +1,6 @@
+package com.teamk.swiths_api.post;
+
+public class PostService {
+
+    
+}
