+++ backend/src/main/java/com/teamk/swiths_api/post/dto/PostRequestDto.java
@@ -1,0 +1,8 @@
+package com.teamk.swiths_api.post.dto;
+
+/**
+ * 
+ */
+public class PostRequestDto {
+    
+}
