+++ backend/src/main/java/com/teamk/swiths_api/post/dto/PostResponseDto.java
@@ -1,0 +1,5 @@
+package com.teamk.swiths_api.post.dto;
+
+public class PostResponseDto {
+    
+}
